Split common entities in a subfolder
