+++ mithril-aggregator/src/http_server/routes/reply.rs
@@ -1,4 +1,4 @@
-use mithril_common::entities::Error;
+use mithril_common::entities::InternalServerError;
 use serde::Serialize;
 use warp::http::StatusCode;
 
@@ -17,5 +17,8 @@ pub fn empty(status_code: StatusCode) -> Box<dyn warp::Reply> {
 }
 
 pub fn internal_server_error(message: String) -> Box<dyn warp::Reply> {
-    json(&Error::new(message), StatusCode::INTERNAL_SERVER_ERROR)
+    json(
+        &InternalServerError::new(message),
+        StatusCode::INTERNAL_SERVER_ERROR,
+    )
 }
