+++ mithril-common/src/entities/snapshot.rs
@@ -1,0 +1,51 @@
+use crate::entities::Beacon;
+use serde::{Deserialize, Serialize};
+
+/// Snapshot represents a snapshot file and its metadata
+#[derive(Clone, Debug, PartialEq, Default, Serialize, Deserialize)]
+pub struct Snapshot {
+    /// Digest that is signed by the signer participants
+    #[serde(rename = "digest")]
+    pub digest: String,
+
+    /// Mithril beacon on the Cardano chain
+    #[serde(rename = "beacon")]
+    pub beacon: Beacon,
+
+    /// Hash of the associated certificate
+    #[serde(rename = "certificate_hash")]
+    pub certificate_hash: String,
+
+    /// Size of the snapshot file in Bytes
+    #[serde(rename = "size")]
+    pub size: u64,
+
+    /// Date and time at which the snapshot was created
+    #[serde(rename = "created_at")]
+    pub created_at: String,
+
+    /// Locations where the binary content of the snapshot can be retrieved
+    #[serde(rename = "locations")]
+    pub locations: Vec<String>,
+}
+
+impl Snapshot {
+    /// Snapshot factory
+    pub fn new(
+        digest: String,
+        beacon: Beacon,
+        certificate_hash: String,
+        size: u64,
+        created_at: String,
+        locations: Vec<String>,
+    ) -> Snapshot {
+        Snapshot {
+            digest,
+            beacon,
+            certificate_hash,
+            size,
+            created_at,
+            locations,
+        }
+    }
+}
